refactor: DeliveryPolicy를 DeliveryValidator로 변경
